+ support providing arbitrary languages param * rename variable `$type` to `$lang` * enable `set -x` for easier debugging @ `protoc.sh`

+++ protoc.sh
@@ -1,27 +1,32 @@
 #!/bin/bash
-PROTO_PATH=tbm_min
-USAGE='Usage: passing following possible args: php cs, to generate corresponding codegen files.'
+set -x
+USAGE='Usage: <PROTO_PATH> <langs...>'
 i=0
-for type in "$@"; do
+for lang in "$@"; do
     i=$((i+1));
-    echo '    codegen for' "$type" 'started:';
-    DIR=${PROTO_PATH}_$type;
+    if [ $i -eq 1 ]; then
+        PROTO_PATH=$lang
+        echo "PROTO_PATH=$PROTO_PATH"
+        continue
+    fi
+
+    echo "    codegen for $lang started:";
+    DIR=${PROTO_PATH}_$lang;
     rm -r "$DIR"
     mkdir -v "$DIR"
-    case $type in
-        php)
-            ./protoc --php_out="$DIR" --proto_path="$PROTO_PATH" "$(find "$PROTO_PATH" -type f -name '*.proto')"
-            ;;
+
+    case $lang in
         cs)
             ./protoc --csharp_out="$DIR" --csharp_opt=base_namespace=TbClient,file_extension=.g.cs --proto_path="$PROTO_PATH" "$(find "$PROTO_PATH" -type f -name '*.proto')"
             ;;
         *)
-            echo "$USAGE"
+            ./protoc --"$lang"_out="$DIR" --proto_path="$PROTO_PATH" "$(find "$PROTO_PATH" -type f -name '*.proto')"
             ;;
     esac
+
     rm -vf "$DIR".7z
     7z a -mx=9 "$DIR".7z ./"$DIR"/*
-    echo '    codegen for' "$type" 'finished.';
+    echo '    codegen for' "$lang" 'finished.';
 done
 
 if [ $i -eq 0 ]; then
